requirements.txt
pymongo[srv]
yfinance
python-dotenv
pandas